Lift QuickInfoSection to its own file and surface in sidebar
Extracts the owner-configured Quick Info tiles (QuickTile type, TileIcon
helper, AddTileModal, EditTileModal, QuickInfoSection) from HomeTab into
src/app/trips/[tripId]/components/QuickInfoSection.tsx. Adds a companion
QuickInfoDrawer — a mobile bottom sheet wrapper — for an upcoming FAB.

SidebarForStage now renders QuickInfoSection for the going/now/past/saved
stages so the trip detail page serves Quick Info from the persistent right
rail instead of HomeTab's own in-tab column, bringing post-planning layout
in line with the planning stage.

HomeTab drops its embedded right column (Quick Info + duplicate Dates
panel) along with the lifted code and now-unused lucide imports.
