+++ src/app/trips/[tripId]/components/SidebarForStage.tsx
@@ -4,6 +4,7 @@ import { MapPin, Plus, Sparkles } from "lucide-react";
 import { SidebarChatPanel } from "./PlanningChatPanel";
 import { StageContextBar } from "./StageContextBar";
 import { CoPlannerPanel } from "./IdeaZonePanel";
+import { QuickInfoSection } from "./QuickInfoSection";
 
 export type SidebarStage = "idea" | "planning" | "going" | "now" | "past" | "saved";
 
@@ -97,6 +98,13 @@ export function SidebarForStage({
           <Sparkles size={15} />
           Trip Summary
         </button>
+      )}
+
+      {/* Post-planning stages (going/now/past/saved) surface owner-configured
+          Quick Info tiles — door codes, check-in times, addresses — as the
+          at-a-glance glance rail alongside chat. */}
+      {(stage === "going" || stage === "now" || stage === "past" || stage === "saved") && (
+        <QuickInfoSection tripId={tripId} isOwner={isOwner} />
       )}
 
       {/* Chat is universal across stages. */}
